+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Possible sections are:
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+- `h3o::is_valid_index` (i.e. `isValidIndex` from H3 4.4.0)
+
 ## [0.9.5] - 2026-05-02
 
 ### Changed